+++ changelog.md
@@ -1,3 +1,12 @@
+### 2021-03-21
+  * Added a bunch of missing game start block info to parser output
+  * Added a bunch of missing player start block info to parser output
+  * Added documentation to all SlippiReplay struct attributes in replay.h
+  * Added original input file name to output of parser and analyzer
+  * Merged cbartsch's fix for item counting
+  * Renamed and reorganized a few fields in output for parser and analyzer
+  * Bumper parser and analyzer versions to 0.7.0
+
 ### 2021-03-13
   * Added new outputs to parser:
     * timer start time (in minutes)
@@ -16,7 +25,6 @@
     * Probably some more tweaks I'm forgetting
   * Enable -O3 optimizations when compiling by default (why were they off?)
   * Merged b3nd3r-ssbm's changes for Windows-compatible endian-swap
-  * Bumper parser and analyzer versions to 0.6.2
 
 ### 2021-02-16
   * Fixed massive heap allocation for SlippiItemFrame wasting a ton of RAM